Refactor Repositories to not use classes with static methods
