+++ backend/medtagger/exceptions.py
@@ -1,0 +1,13 @@
+"""All available Exceptions for whole project."""
+
+
+class UnsupportedActionException(Exception):
+    """Exception for unsupported Action."""
+
+    pass
+
+
+class InvalidResponseException(Exception):
+    """Exception for invalid Response."""
+
+    pass
