+++ backend/medtagger/repositories/scan_categories.py
@@ -1,63 +1,59 @@
-"""Module responsible for definition of Scan Categories' Repository."""
+"""Module responsible for definition of ScanCategoriesRepository."""
 from typing import List
 
 from medtagger.database import db_session
 from medtagger.database.models import ScanCategory, LabelTag
 
 
-class ScanCategoriesRepository(object):
-    """Repository for Scan Categories."""
+def get_all_categories() -> List[ScanCategory]:
+    """Return list of all Scan Categories."""
+    with db_session() as session:
+        categories = session.query(ScanCategory).order_by(ScanCategory.id).all()
+    return categories
 
-    @staticmethod
-    def get_all_categories() -> List[ScanCategory]:
-        """Return list of all Scan Categories."""
-        with db_session() as session:
-            categories = session.query(ScanCategory).order_by(ScanCategory.id).all()
-        return categories
 
-    @staticmethod
-    def get_category_by_key(key: str) -> ScanCategory:
-        """Fetch Scan Category from database.
+def get_category_by_key(key: str) -> ScanCategory:
+    """Fetch Scan Category from database.
 
-        :param key: key for a Scan Category
-        :return: Scan Category object
-        """
-        with db_session() as session:
-            category = session.query(ScanCategory).filter(ScanCategory.key == key).one()
-        return category
+    :param key: key for a Scan Category
+    :return: Scan Category object
+    """
+    with db_session() as session:
+        category = session.query(ScanCategory).filter(ScanCategory.key == key).one()
+    return category
 
-    @staticmethod
-    def add_new_category(key: str, name: str, image_path: str) -> ScanCategory:
-        """Add new Scan Category to the database.
 
-        :param key: key that will identify such Scan Category
-        :param name: name that will be used in the Use Interface for such Scan Category
-        :param image_path: path to the image that represents such Scan Category (used in User Interface)
-        :return: Scan Category object
-        """
-        with db_session() as session:
-            category = ScanCategory(key, name, image_path)
-            session.add(category)
-        return category
+def add_new_category(key: str, name: str, image_path: str) -> ScanCategory:
+    """Add new Scan Category to the database.
 
-    @staticmethod
-    def assign_label_tag(tag: LabelTag, scan_category_key: str) -> None:
-        """Assign existing Label Tag to Scan Category.
+    :param key: key that will identify such Scan Category
+    :param name: name that will be used in the Use Interface for such Scan Category
+    :param image_path: path to the image that represents such Scan Category (used in User Interface)
+    :return: Scan Category object
+    """
+    with db_session() as session:
+        category = ScanCategory(key, name, image_path)
+        session.add(category)
+    return category
 
-        :param tag: tag that should be assigned to Scan Category
-        :param scan_category_key: key that will identify such Scan Category
-        """
-        scan_category = ScanCategory.query.filter(ScanCategory.key == scan_category_key).one()
-        scan_category.available_tags.append(tag)
-        scan_category.save()
 
-    @staticmethod
-    def unassign_label_tag(tag: LabelTag, scan_category_key: str) -> None:
-        """Unassign Label Tag from Scan Category.
+def assign_label_tag(tag: LabelTag, scan_category_key: str) -> None:
+    """Assign existing Label Tag to Scan Category.
 
-        :param tag: tag that should be unassigned from Scan Category
-        :param scan_category_key: key that will identify such Scan Category
-        """
-        scan_category = ScanCategory.query.filter(ScanCategory.key == scan_category_key).one()
-        scan_category.available_tags.remove(tag)
-        scan_category.save()
+    :param tag: tag that should be assigned to Scan Category
+    :param scan_category_key: key that will identify such Scan Category
+    """
+    scan_category = ScanCategory.query.filter(ScanCategory.key == scan_category_key).one()
+    scan_category.available_tags.append(tag)
+    scan_category.save()
+
+
+def unassign_label_tag(tag: LabelTag, scan_category_key: str) -> None:
+    """Unassign Label Tag from Scan Category.
+
+    :param tag: tag that should be unassigned from Scan Category
+    :param scan_category_key: key that will identify such Scan Category
+    """
+    scan_category = ScanCategory.query.filter(ScanCategory.key == scan_category_key).one()
+    scan_category.available_tags.remove(tag)
+    scan_category.save()
