+++ backend/tests/functional_tests/helpers.py
@@ -1,7 +1,6 @@
 """Helpers functions for functional tests."""
 from medtagger.database.models import LabelTag
-from medtagger.repositories.label_tag import LabelTagRepository
-from medtagger.repositories.scan_categories import ScanCategoriesRepository
+from medtagger.repositories import label_tags as LabelTagsRepository, scan_categories as ScanCategoriesRepository
 
 
 def create_tag_and_assign_to_category(key: str, name: str, scan_category_key: str) -> LabelTag:
@@ -12,6 +11,6 @@ def create_tag_and_assign_to_category(key: str, name: str, scan_category_key: st
     :param scan_category_key: key of the Scan Category that Label Tag will be assigned to
     :return: Label Tag
     """
-    label_tag = LabelTagRepository.add_new_tag(key, name)
+    label_tag = LabelTagsRepository.add_new_tag(key, name)
     ScanCategoriesRepository.assign_label_tag(label_tag, scan_category_key)
     return label_tag
